Various changes
Better MS5541C init messages
Pressure sensor with Robot tested
Added SPI.h to Cubeception.ino

+++ Cubeception/src/RobotController/Robot/Robot.h
@@ -209,7 +209,7 @@ private:
     
     //For pressure sensor
     bool _temp;
-    int8_t _queueTime;
+    double _queueTime;
     unsigned long _timeSinceQueuing;
     
     //Sensors and actuators
